Moved Wii save format stuff into its own directory
